// src/app/client/model/marital-status.enum.ts
export enum MaritalStatus {
  UNKNOWN= 'unknown',
  SINGLE = 'Ledig',
  MARRIED = 'Verheiratet',
  DIVORCED = 'Geschieden',
  WIDOWED = 'Verwitwet',
  REGISTERED_PARTNERSHIP = 'Eingetragene Partnerschaft',
  DISSOLVED_REGISTERED_PARTNERSHIP = 'Aufgelöste eingetragene Partnerschaft',
  SURVIVING_REGISTERED_PARTNERSHIP = 'Hinterbliebener eingetragene Partnerschaft'
}